git: pull merge instead of rebase
tldr: avoid unexpected conflict during git pull

pull rebase was originally set so that when the local and origin
branches diverge (eg: when two or more people are working on the same
branch) commits would be rebased into a single linear branch.

however, when the local branch is ahead of the origin, and contains a
merge commit with conflicts, eg:

*   18b3fcd - (HEAD -> feature2) Merge branch 'fixes' into feature2
|\
| * 910ad33 - (origin/fixes, fixes) fixed!
* | 85d91af - (origin/feature2) wip
|/
* 67b2593 - first feature

then git pull --rebase run from HEAD will attempt to linearise the
branch by replaying all commits from the common ancestor (67b2593).
This reintroduces the conflict, which will need to be resolved again.

The conflict is surprising because the local branch is ahead of the
remote branch, and so there can't be anything on the remote branch
that is conflicting.

git pull --rebase=merges will keep the local merge commit and avoid the
conflict. git pull --no-rebase will do a merge instead of a rebase and
also avoid the conflict. This will be a fast-forward if possible,
otherwise a merge commit is created which breaks linearity but at is
explicit. Since no rebase is also the git default, and keeping the
branch linear doesn't matter much when squash merging into main (and
the branch is deleted) let's switch back to that to avoid the conflict
case above.